// src/types/compliance.ts
export interface IFRSFinancialStatements {
  statementOfProfitOrLoss: IFRSProfitLoss;
  statementOfFinancialPosition: IFRSBalanceSheet;
  statementOfCashFlows: IFRSCashFlow;
  statementOfChangesInEquity: IFRSEquity;
  notesToFinancialStatements: IFRSNotes[];
  comparativeFigures: ComparativeFigures;
}

export interface IFRSProfitLoss {
  revenue: RevenueLine[];
  costOfSales: CostOfSalesLine[];
  grossProfit: number;
  otherIncome: OtherIncomeLine[];
  operatingExpenses: OperatingExpense[];
  profitBeforeTax: number;
  taxExpense: TaxExpenseLine[];
  profitAfterTax: number;
  otherComprehensiveIncome: OtherComprehensiveIncome[];
  totalComprehensiveIncome: number;
  earningsPerShare: EarningsPerShare;
}

export interface IFRSBalanceSheet {
  assets: AssetSection[];
  liabilities: LiabilitySection[];
  equity: EquitySection[];
  totalAssets: number;
  totalLiabilitiesAndEquity: number;
}

export interface IFRSCashFlow {
  operatingActivities: CashFlowItem[];
  investingActivities: CashFlowItem[];
  financingActivities: CashFlowItem[];
  netIncreaseInCash: number;
  cashAtBeginningOfPeriod: number;
  cashAtEndOfPeriod: number;
}

export interface IFRSEquity {
  shareCapital: ShareCapital;
  sharePremium: SharePremium;
  retainedEarnings: RetainedEarnings;
  otherReserves: OtherReserve[];
  totalEquity: number;
}

export interface IFRSNotes {
  accountingPolicies: AccountingPolicies[];
  segmentInformation: SegmentInformation[];
  relatedPartyTransactions: RelatedPartyTransaction[];
  contingencies: Contingency[];
  subsequentEvents: SubsequentEvent[];
  fairValueMeasurements: FairValueMeasurement[];
}